refactor: remove support to entity Usuario
