--- restaurante-application-core/src/main/java/br/com/nomeempresa/restaurante/core/domain/entities/Usuario.java
@@ -1,45 +1,0 @@
-package br.com.nomeempresa.restaurante.core.domain.entities;
-
-import java.io.Serializable;
-
-public class Usuario implements Serializable {
-
-    private Long id;
-
-    private String nome;
-
-    private String cpf;
-
-    private String email;
-
-    public Long getId() {
-        return id;
-    }
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
